+++ samples/DotNetCoreSample/Program.cs
@@ -41,6 +41,7 @@ namespace DotNetCoreSample
             //}
 
             var list2 = new List<TestEntity2>();
+            list2.Add(null);
             for (var i = 0; i < 10; i++)
             {
                 list2.Add(new TestEntity2
@@ -52,6 +53,8 @@ namespace DotNetCoreSample
             }
             var tempDirPath = $@"{Environment.GetEnvironmentVariable("USERPROFILE")}\Desktop\temp\test";
             list2.ToExcelFile($@"{tempDirPath}\testEntity2.xlsx");
+
+            var listTemp = ExcelHelper.ToEntityList<TestEntity2>($@"{tempDirPath}\testEntity2.xlsx");
 
             Console.WriteLine("Press Enter to continue...");
             Console.ReadLine();